move components into views
