+++ client/src/views/CreateSchedule.js
@@ -1,13 +1,13 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import Scheduler from './Scheduler'
-import Modal from './Modal'
-import HoursOfOperationForm from './HoursOfOperationForm'
-import TopBar from './common/TopBar'
-import NavMenu from './common/NavMenu'
-import OuterContainer from './common/OuterContainer'
+import Scheduler from '../components/Scheduler'
+import Modal from '../components/Modal'
+import HoursOfOperationForm from '../components/HoursOfOperationForm'
+import TopBar from '../components/common/TopBar'
+import NavMenu from '../components/common/NavMenu'
+import OuterContainer from '../components/common/OuterContainer'
 import styled from '@emotion/styled'
-import Alert from './common/Alert'
+import Alert from '../components/common/Alert'
 import system from '../design/theme'
 
 // this component will house all of the main features for the create schedule page.
