+++ client/src/views/Employees.js
@@ -1,20 +1,20 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import TopBar from './common/TopBar'
+import TopBar from '../components/common/TopBar'
 import styled from '@emotion/styled'
 import system from '../design/theme'
 import { fetchEmployeesFromDB, updateUserSettings } from '../actions'
 import { connect } from 'react-redux'
-import Card from './EmployeeCard/Card'
-import NavMenu from './common/NavMenu'
-import OuterContainer from './common/OuterContainer'
-import AddEmployee from './AddEmployee'
-import AvailabilityForm from './Availability/AvailabilityForm'
-import Button from './common/Button'
-import Modal from './Modal'
-import { Input } from './common/FormContainer'
+import Card from '../components/EmployeeCard/Card'
+import NavMenu from '../components/common/NavMenu'
+import OuterContainer from '../components/common/OuterContainer'
+import AddEmployee from '../components/AddEmployee'
+import AvailabilityForm from '../components/Availability/AvailabilityForm'
+import Button from '../components/common/Button'
+import Modal from '../components/Modal'
+import { Input } from '../components/common/FormContainer'
 import ReactJoyride, { STATUS, ACTIONS } from 'react-joyride'
-import steps from './Demo/employees'
+import steps from '../components/Demo/employees'
 import axios from 'axios'
 const baseURL = process.env.REACT_APP_SERVER_URL
