+++ client/src/views/Home.js
@@ -1,14 +1,14 @@
 import React from 'react'
-import TopBar from './common/TopBar'
+import TopBar from '../components/common/TopBar'
 import styled from '@emotion/styled'
 import system from '../design/theme'
-import Button from './common/Button'
-import LinkItem from './common/LinkItem'
-import OuterContainer from './common/OuterContainer'
+import Button from '../components/common/Button'
+import LinkItem from '../components/common/LinkItem'
+import OuterContainer from '../components/common/OuterContainer'
 import headerImg from '../img/header.svg'
 import heroImg from '../img/hero.svg'
 import Fade from 'react-reveal/Fade'
-import Footer from './common/Footer'
+import Footer from '../components/common/Footer'
 import frontpage from '../img/frontpage.mp4'
 
 const Home = () => {
