+++ client/src/views/RegisterOwner.js
@@ -13,14 +13,14 @@ import {
   logout
 } from '../actions' // for calling once all data is in
 import { connect } from 'react-redux'
-import Login from './Login'
-import Status from './common/Status'
-import LayoutFrame from './common/LayoutFrame'
-import TopBar from './common/TopBar'
-import OuterContainer from './common/OuterContainer'
-import { Container, Input, Select } from './common/FormContainer'
+import Login from '../components/Login'
+import Status from '../components/common/Status'
+import LayoutFrame from '../components/common/LayoutFrame'
+import TopBar from '../components/common/TopBar'
+import OuterContainer from '../components/common/OuterContainer'
+import { Container, Input, Select } from '../components/common/FormContainer'
 import { Redirect } from 'react-router-dom'
-import Button from './common/Button'
+import Button from '../components/common/Button'
 import industryList from '../industryList'
 import styled from '@emotion/styled'
 import moment from 'moment'
