+++ client/src/views/Settings.js
@@ -1,14 +1,14 @@
 import React, { Component } from 'react'
 import propTypes from 'prop-types'
-import TopBar from './common/TopBar'
-import NavMenu from './common/NavMenu'
-import OuterContainer from './common/OuterContainer'
-import { Container, Input } from './common/FormContainer'
-import Button from './common/Button'
+import TopBar from '../components/common/TopBar'
+import NavMenu from '../components/common/NavMenu'
+import OuterContainer from '../components/common/OuterContainer'
+import { Container, Input } from '../components/common/FormContainer'
+import Button from '../components/common/Button'
 import axios from 'axios'
 import { connect } from 'react-redux'
-import Status from './common/Status'
-import Loader from './common/Loader'
+import Status from '../components/common/Status'
+import Loader from '../components/common/Loader'
 import { updateUserSettings } from '../actions'
 
 const phonePattern =
