+++ client/src/views/Team.js
@@ -1,7 +1,7 @@
 import React from 'react'
-import OuterContainer from './common/OuterContainer'
-import TopBar from './common/TopBar'
-import Footer from './common/Footer'
+import OuterContainer from '../components/common/OuterContainer'
+import TopBar from '../components/common/TopBar'
+import Footer from '../components/common/Footer'
 import styled from '@emotion/styled'
 import system from '../design/theme'
 import Kam from '../img/Kam.jpg'
